+++ koin-projects/examples/androidx-compose-jetnews/src/main/java/com/example/jetnews/ui/JetnewsApp.kt
@@ -45,33 +45,30 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import androidx.ui.tooling.preview.Preview
 import com.example.jetnews.R
-import com.example.jetnews.data.AppContainer
 import com.example.jetnews.data.interests.InterestsRepository
 import com.example.jetnews.data.posts.PostsRepository
 import com.example.jetnews.ui.article.ArticleScreen
 import com.example.jetnews.ui.home.HomeScreen
 import com.example.jetnews.ui.interests.InterestsScreen
 import com.example.jetnews.ui.theme.JetnewsTheme
+import org.koin.androidx.compose.get
 
 @Composable
 fun JetnewsApp(
-    appContainer: AppContainer,
     navigationViewModel: NavigationViewModel
 ) {
     JetnewsTheme {
         AppContent(
-            navigationViewModel = navigationViewModel,
-            interestsRepository = appContainer.interestsRepository,
-            postsRepository = appContainer.postsRepository
+            navigationViewModel = navigationViewModel
         )
     }
 }
 
 @Composable
 private fun AppContent(
-    navigationViewModel: NavigationViewModel,
-    postsRepository: PostsRepository,
-    interestsRepository: InterestsRepository
+        navigationViewModel: NavigationViewModel,
+        postsRepository: PostsRepository = get(),
+        interestsRepository: InterestsRepository = get()
 ) {
     Crossfade(navigationViewModel.currentScreen) { screen ->
         Surface(color = MaterialTheme.colors.background) {